Separate `hostdevice_vector` out from `gdf_column_wrapper` header
